Move PlayButtons and CancelButton into another file.
PlayControls ended up being much large than I thought, and I plan to move
even more state out of Game and into these components, so breaking out the
PlayButtons seems like a good idea.
